Fixing the issue with incorrect resource packaging in bit6 (the previous solution did not work when building-packaging bit6 from scratch).
The hook script is changed to run "ant release" after plugin install to have proper configuration.
(the old hook logic to modify project.properties is not needed, since using <framework> tag in plugin.xml does the thing).

TODO: this is not the best solution, so may need refatoring later. Pushing this to move forward.